==> 2_data_mining.ipynb <==
# Data Mining

## Import Libraries

from datetime import datetime
import pandas as pd
import glob
import numpy as np
from scripts.helper import reduce_mem_usage, load_df

## Import Datasets

# Set path to train and test datasets
path = 'data/train_v2.csv'
path1 = 'data/test_v2.csv'

# Define JSON columns to extract
json_columns = ['device', 'geoNetwork', 'totals', 'trafficSource']

# Define features to extract
features = ['channelGrouping', 'date', 'fullVisitorId', 'visitId',\
       'visitNumber', 'visitStartTime', 'device.browser',\
       'device.deviceCategory', 'device.isMobile', 'device.operatingSystem',\
       'geoNetwork.city', 'geoNetwork.continent', 'geoNetwork.country',\
       'geoNetwork.metro', 'geoNetwork.networkDomain', 'geoNetwork.region',\
       'geoNetwork.subContinent', 'totals.bounces', 'totals.hits',\
       'totals.newVisits', 'totals.pageviews', 'totals.transactionRevenue',\
       'trafficSource.adContent', 'trafficSource.campaign',\
       'trafficSource.isTrueDirect', 'trafficSource.keyword',\
       'trafficSource.medium', 'trafficSource.referralPath',\
       'trafficSource.source']

start_time = datetime.now() # Set start point for time analysis

# Extract the features from the JSON columns
dataset = load_df(path, json_columns, features)
dataset_test = load_df(path1, json_columns, features)

end_time = datetime.now() # Set end point for time analysis
print('Duration: {}'.format(end_time - start_time)) # Print out anlysed time

## Shrink data

# Reduce dataset size by change the datatypes
dataset, NAlist = reduce_mem_usage(dataset)
dataset_test, NAlist = reduce_mem_usage(dataset_test)

# Show head for train dataset
dataset.head()

# Show head for test dataset
dataset_test.head()

## Export Datasets

# Save shrunk train data to a new .csv file
path = 'data/shrunk_train_v2.csv'
dataset.to_csv(path)

# Save shrunk test data to a new .csv file
path1 = 'data/shrunk_test_v2.csv'
dataset_test.to_csv(path1)
